Server thread is functional, should be final version. Confirmed that the queue can still be accessed by the main process.
