+++ misc/udp_traffictest.py
@@ -5,24 +5,17 @@ import time
 bufferSize  = 1024
 serverAddressPort   = ("127.0.0.1", 7500)
 
-
 print('this program will generate some test traffic for 2 players on the red ')
 print('team as well as 2 players on the green team')
 print('')
 
-#red1 = input('Enter id of red player 1 ==> ')
-#red2 = input('Enter id of red player 2 ==> ')
-#green1 = input('Enter id of green player 1 ==> ')
-#green2 = input('Enter id of green player 2 ==> ')
+red1 = input('Enter id of red player 1 ==> ')
+red2 = input('Enter id of red player 2 ==> ')
+green1 = input('Enter id of green player 1 ==> ')
+green2 = input('Enter id of green player 2 ==> ')
 
-red1 = '1'
-red2 = '2'
-green1 = '3'
-green2 = '4'
-
-#print('')
-#counter = input('How many events do you want ==> ')
-counter = 20
+print('')
+counter = input('How many events do you want ==> ')
 
 # Create datagram socket
 UDPClientSocketTransmit = socket.socket(family=socket.AF_INET, type=socket.SOCK_DGRAM)
